refactor: change Album to AlbumMain
